fix docker build
