+++ crypto/keyring/options_notavail.go
@@ -1,0 +1,17 @@
+//go:build !cgo || !ledger
+// +build !cgo !ledger
+
+package keyring
+
+import (
+	"github.com/cosmos/cosmos-sdk/crypto/keyring"
+)
+
+// EthSecp256k1Option defines a function keys options for the ethereum Secp256k1 curve.
+// It supports eth_secp256k1 keys for accounts.
+func EthSecp256k1Option() keyring.Option {
+	return func(options *keyring.Options) {
+		options.SupportedAlgos = SupportedAlgorithms
+		options.SupportedAlgosLedger = SupportedAlgorithmsLedger
+	}
+}
