+++ crypto/ledger/ledger_notavail.go
@@ -1,8 +1,6 @@
 //go:build !cgo || !ledger
 // +build !cgo !ledger
 
-// test_ledger_mock
-
 package ledger
 
 import (
